rm duplicates and useless imports
